+++ api/app/controllers/v1/users_controller.rb
@@ -1,0 +1,25 @@
+module V1
+    class UsersController < ApplicationController
+
+        def new
+            @user = User.new
+        end
+        
+        def create
+            @user = User.new user_params
+            if @user.save
+                session[:user_id] = @user.id
+            redirect_to root_path, notice: "Logged In!"
+            else
+                render :new
+            end
+        end
+        private
+        
+        def user_params
+            params.require(:user).permit(:first_name, :last_name, :email,
+            :password, :password_confirmation)
+        end
+
+    end
+end
